+++ src/util.h
@@ -471,7 +471,7 @@ QCString getEncoding(const FileInfo &fi);
 
 inline QCString fixSpaces(const QCString &s) { return substitute(s," ","&#160;"); }
 
-QCString detab(const QCString &s,int &refIndent);
+QCString detab(const QCString &s,size_t &refIndent);
 
 QCString getProjectId();
 QCString projectLogoFile();
